# Add `-perslice` option (plus performance improvements to make it feasible)