--- a/spinalcordtoolbox/scripts/sct_analyze_lesion.py
+++ b/spinalcordtoolbox/scripts/sct_analyze_lesion.py
@@ -528,7 +528,14 @@ class AnalyzeLesion:
     def ___pve_weighted_avg(self, im_mask_data, im_atlas_data):
         return im_mask_data * im_atlas_data
 
-    def __relative_ROIvol_in_mask(self, im_mask_data, im_atlas_roi_data, p_lst, im_template_vert_data=None, vert_level=None):
+    def __keep_only_indices(self, image, indices):
+        """Keep values defined by indices, and set all other coordinates to zero."""
+        image_out = np.zeros_like(image)
+        image_out[indices] = image[indices]
+        return image_out
+
+    def __relative_ROIvol_in_mask(self, im_mask_data, im_atlas_roi_data, p_lst, im_template_vert_data=None,
+                                  vert_level=None, vert_indices=None):
         #
         #   Goal:
         #         This function computes the percentage of ROI occupied by binary mask
@@ -544,9 +551,9 @@ class AnalyzeLesion:
         #           - vert_level - type=int - vertebral level ID to restrict the ROI
         #
 
-        if im_template_vert_data is not None and vert_level is not None:
-            im_atlas_roi_data[np.where(im_template_vert_data != vert_level)] = 0.0
-            im_mask_data[np.where(im_template_vert_data != vert_level)] = 0.0
+        if im_template_vert_data is not None and vert_level in self.vert_lst:
+            im_atlas_roi_data = self.__keep_only_indices(im_atlas_roi_data, vert_indices[vert_level])
+            im_mask_data = self.__keep_only_indices(im_mask_data, vert_indices[vert_level])
 
         im_mask_roi_data_wa = self.___pve_weighted_avg(im_mask_data=im_mask_data, im_atlas_data=im_atlas_roi_data)
         vol_tot_roi = np.sum(im_atlas_roi_data) * p_lst[0] * p_lst[1] * p_lst[2]
@@ -554,7 +561,7 @@ class AnalyzeLesion:
 
         return vol_mask_roi_wa, vol_tot_roi
 
-    def _measure_eachLesion_distribution(self, lesion_id, atlas_data, im_vert, im_lesion, p_lst):
+    def _measure_eachLesion_distribution(self, lesion_id, atlas_data, im_vert, im_lesion, p_lst, vert_indices):
         sheet_name = 'lesion#' + str(lesion_id) + '_distribution'
         self.distrib_matrix_dct[sheet_name] = pd.DataFrame.from_dict({'vert': [str(v) for v in self.vert_lst]})
 
@@ -564,8 +571,8 @@ class AnalyzeLesion:
 
         vol_mask_tot = 0.0  # vol tot of this lesion through the vertebral levels and PAM50 tracts
         for vert in self.vert_lst:  # Loop over vertebral levels
-            im_vert_cur = np.copy(im_vert)
-            im_vert_cur[np.where(im_vert_cur != vert)] = 0.0
+            print(f"  Computing lesion distribution for vert {vert}...")
+            im_vert_cur = self.__keep_only_indices(im_vert, vert_indices[vert])
             if np.count_nonzero(im_vert_cur * np.copy(im_lesion)):  # if there is lesion in this vertebral level
                 idx = self.distrib_matrix_dct[sheet_name][self.distrib_matrix_dct[sheet_name].vert == str(vert)].index
                 for tract_id in atlas_data:  # Loop over PAM50 tracts
@@ -573,7 +580,8 @@ class AnalyzeLesion:
                                                              im_atlas_roi_data=np.copy(atlas_data[tract_id]),
                                                              p_lst=p_lst,
                                                              im_template_vert_data=np.copy(im_vert_cur),
-                                                             vert_level=vert)
+                                                             vert_level=vert,
+                                                             vert_indices=vert_indices)
                     self.distrib_matrix_dct[sheet_name].loc[idx, 'PAM50_' + str(tract_id).zfill(2)] = res_lst[0]
                     vol_mask_tot += res_lst[0]
 
@@ -608,7 +616,7 @@ class AnalyzeLesion:
         res_tot = [vol_dct[t][1] for t in vol_dct if t in tracts]
         return np.sum(res_mask) * 100.0 / np.sum(res_tot)
 
-    def _measure_totLesion_distribution(self, im_lesion, atlas_data, im_vert, p_lst):
+    def _measure_totLesion_distribution(self, im_lesion, atlas_data, im_vert, p_lst, vert_indices):
 
         sheet_name = 'ROI_occupied_by_lesion'
         self.distrib_matrix_dct[sheet_name] = pd.DataFrame.from_dict({'vert': [str(v) for v in self.vert_lst] + ['total % (all vert)']})
@@ -618,9 +626,9 @@ class AnalyzeLesion:
             self.distrib_matrix_dct[sheet_name]['PAM50_' + str(tract_id).zfill(2)] = [0] * len(self.vert_lst + ['total % (all vert)'])
 
         for vert in self.vert_lst + ['total % (all vert)']:  # loop over the vertebral levels
+            print(f"  Computing ROI distribution for vert {vert}...")
             if vert != 'total % (all vert)':
-                im_vert_cur = np.copy(im_vert)
-                im_vert_cur[np.where(im_vert_cur != vert)] = 0
+                im_vert_cur = self.__keep_only_indices(im_vert, vert_indices[vert])
             else:
                 im_vert_cur = None
             if im_vert_cur is None or np.count_nonzero(im_vert_cur * np.copy(im_lesion)):
@@ -631,7 +639,8 @@ class AnalyzeLesion:
                                                                                 im_atlas_roi_data=np.copy(atlas_data[tract_id]),
                                                                                 p_lst=p_lst,
                                                                                 im_template_vert_data=np.copy(im_vert_cur),
-                                                                                vert_level=vert)
+                                                                                vert_level=vert,
+                                                                                vert_indices=vert_indices)
 
                 # group tracts to compute involvement in CombinedLabels (GM, WM, DC, VF, LF)
                 for label_name, sublabels in self.atlas_combinedlabels.items():
@@ -658,6 +667,11 @@ class AnalyzeLesion:
                 img_vert = Image(self.path_levels)
                 im_vert_data = img_vert.data
                 self.vert_lst = [v for v in np.unique(im_vert_data) if v]  # list of vertebral levels available in the input image
+                # precompute the list of indices for each vertebral level
+                # these indices are used to zero out certain levels (to compute the volume of the remaining levels)
+                vert_indices = {
+                    vert: np.where(im_vert_data == vert) for vert in self.vert_lst
+                }
 
             else:
                 im_vert_data = None
@@ -701,14 +715,17 @@ class AnalyzeLesion:
                                                       atlas_data=atlas_data_dct,
                                                       im_vert=im_vert_data,
                                                       im_lesion=im_lesion_data_cur,
-                                                      p_lst=p_lst)
+                                                      p_lst=p_lst,
+                                                      vert_indices=vert_indices)
 
         if self.path_template is not None:
             # compute total lesion distribution
+            print("\nROI percentage taken up by all lesions...")
             self._measure_totLesion_distribution(im_lesion=np.copy(im_lesion_data > 0),
                                                  atlas_data=atlas_data_dct,
                                                  im_vert=im_vert_data,
-                                                 p_lst=p_lst)
+                                                 p_lst=p_lst,
+                                                 vert_indices=vert_indices)
 
         if self.fname_ref is not None:
             # Compute mean and std value in each labeled lesion