--- a/spinalcordtoolbox/scripts/sct_analyze_lesion.py
+++ b/spinalcordtoolbox/scripts/sct_analyze_lesion.py
@@ -17,7 +17,7 @@ from spinalcordtoolbox.image import Image
 from spinalcordtoolbox.centerline.core import ParamCenterline, get_centerline
 from spinalcordtoolbox.metadata import read_label_file
 from spinalcordtoolbox.utils.shell import SCTArgumentParser, Metavar, ActionCreateFolder, display_viewer_syntax
-from spinalcordtoolbox.utils.sys import init_sct, printv, set_loglevel, LazyLoader
+from spinalcordtoolbox.utils.sys import init_sct, printv, set_loglevel, LazyLoader, sct_progress_bar
 from spinalcordtoolbox.utils.fs import tmp_create, extract_fname, copy, rmtree
 from spinalcordtoolbox.reports.qc2 import sct_analyze_lesion
 
@@ -88,6 +88,16 @@ def get_parser():
         default=None,
         required=False)
     optional.add_argument(
+        "-perslice",
+        help="Specify whether to aggregate atlas metrics ('-f' option) per slice (`-perslice 1`) or per vertebral "
+             "level (default behavior).",
+        metavar=Metavar.int,
+        type=int,
+        choices=(0, 1),
+        default=0,
+        required=False
+    )
+    optional.add_argument(
         "-ofolder",
         help='Output folder (e.g. "."). Default is the current folder (".").',
         metavar=Metavar.folder,
@@ -130,7 +140,7 @@ def get_parser():
 
 
 class AnalyzeLesion:
-    def __init__(self, fname_mask, fname_sc, fname_ref, path_template, path_ofolder, verbose):
+    def __init__(self, fname_mask, fname_sc, fname_ref, path_template, path_ofolder, perslice, verbose):
         self.fname_mask = fname_mask
 
         self.fname_sc = fname_sc
@@ -177,7 +187,8 @@ class AnalyzeLesion:
             self.path_levels = os.path.join(self.path_template, "template", "PAM50_levels.nii.gz")
         else:
             self.path_atlas, self.path_levels = None, None
-        self.vert_lst = None
+        self.rows = None
+        self.row_name = "slice" if perslice else "vert"
         self.atlas_roi_lst = None
         self.atlas_combinedlabels = {}
         self.distrib_matrix_dct = {}
@@ -531,11 +542,14 @@ class AnalyzeLesion:
     def __keep_only_indices(self, image, indices):
         """Keep values defined by indices, and set all other coordinates to zero."""
         image_out = np.zeros_like(image)
-        image_out[indices] = image[indices]
+        if isinstance(indices, list) and len(indices) == 3:  # Lists of [x], [y], and [z] coords produced by np.where()
+            image_out[indices] = image[indices]
+        else:
+            image_out[:, :, indices] = image[:, :, indices]
         return image_out
 
     def __relative_ROIvol_in_mask(self, im_mask_data, im_atlas_roi_data, p_lst, im_template_vert_data=None,
-                                  vert_level=None, vert_indices=None):
+                                  indices_to_keep=None):
         #
         #   Goal:
         #         This function computes the percentage of ROI occupied by binary mask
@@ -550,10 +564,9 @@ class AnalyzeLesion:
         #           - im_template_vert_data - type=NumPyArray - vertebral template in the same space as im_mask
         #           - vert_level - type=int - vertebral level ID to restrict the ROI
         #
-
-        if im_template_vert_data is not None and vert_level in self.vert_lst:
-            im_atlas_roi_data = self.__keep_only_indices(im_atlas_roi_data, vert_indices[vert_level])
-            im_mask_data = self.__keep_only_indices(im_mask_data, vert_indices[vert_level])
+        if im_template_vert_data is not None and indices_to_keep:
+            im_atlas_roi_data = self.__keep_only_indices(im_atlas_roi_data, indices_to_keep)
+            im_mask_data = self.__keep_only_indices(im_mask_data, indices_to_keep)
 
         im_mask_roi_data_wa = self.___pve_weighted_avg(im_mask_data=im_mask_data, im_atlas_data=im_atlas_roi_data)
         vol_tot_roi = np.sum(im_atlas_roi_data) * p_lst[0] * p_lst[1] * p_lst[2]
@@ -563,31 +576,33 @@ class AnalyzeLesion:
 
     def _measure_eachLesion_distribution(self, lesion_id, atlas_data, im_vert, im_lesion, p_lst, vert_indices):
         sheet_name = 'lesion#' + str(lesion_id) + '_distribution'
-        self.distrib_matrix_dct[sheet_name] = pd.DataFrame.from_dict({'vert': [str(v) for v in self.vert_lst]})
+        self.distrib_matrix_dct[sheet_name] = pd.DataFrame.from_dict({'row': [str(v) for v in self.rows]})
 
         # initialized to 0 for each vertebral level and each PAM50 tract
         for tract_id in atlas_data:
-            self.distrib_matrix_dct[sheet_name]['PAM50_' + str(tract_id).zfill(2)] = [0] * len(self.vert_lst)
+            self.distrib_matrix_dct[sheet_name]['PAM50_' + str(tract_id).zfill(2)] = [0] * len(self.rows)
 
         vol_mask_tot = 0.0  # vol tot of this lesion through the vertebral levels and PAM50 tracts
-        for vert in self.vert_lst:  # Loop over vertebral levels
-            print(f"  Computing lesion distribution for vert {vert}...")
-            im_vert_cur = self.__keep_only_indices(im_vert, vert_indices[vert])
+        for row in sct_progress_bar(self.rows, unit=self.row_name,
+                                    desc="  Computing lesion distribution (volume values)"):  # Loop over slices or vertebral levels
+            # Set everything but the current vertlevel (or slice) to 0
+            indices_to_keep = vert_indices[row] if self.row_name == "vert" else row
+            im_vert_cur = self.__keep_only_indices(im_vert, indices_to_keep)
             if np.count_nonzero(im_vert_cur * np.copy(im_lesion)):  # if there is lesion in this vertebral level
-                idx = self.distrib_matrix_dct[sheet_name][self.distrib_matrix_dct[sheet_name].vert == str(vert)].index
+                idx = self.distrib_matrix_dct[sheet_name][self.distrib_matrix_dct[sheet_name].row == str(row)].index
                 for tract_id in atlas_data:  # Loop over PAM50 tracts
                     res_lst = self.__relative_ROIvol_in_mask(im_mask_data=np.copy(im_lesion),
                                                              im_atlas_roi_data=np.copy(atlas_data[tract_id]),
                                                              p_lst=p_lst,
                                                              im_template_vert_data=np.copy(im_vert_cur),
-                                                             vert_level=vert,
-                                                             vert_indices=vert_indices)
+                                                             indices_to_keep=indices_to_keep)
                     self.distrib_matrix_dct[sheet_name].loc[idx, 'PAM50_' + str(tract_id).zfill(2)] = res_lst[0]
                     vol_mask_tot += res_lst[0]
 
         # convert the volume values in distrib_matrix_dct to percentage values
-        for vert in self.vert_lst:
-            idx = self.distrib_matrix_dct[sheet_name][self.distrib_matrix_dct[sheet_name].vert == str(vert)].index
+        for row in sct_progress_bar(self.rows, unit=self.row_name,
+                                    desc="  Converting volume values into percentage values"):
+            idx = self.distrib_matrix_dct[sheet_name][self.distrib_matrix_dct[sheet_name].row == str(row)].index
             for tract_id in atlas_data:
                 val = self.distrib_matrix_dct[sheet_name].loc[idx, 'PAM50_' + str(tract_id).zfill(2)].values[0]
                 self.distrib_matrix_dct[sheet_name].loc[idx, 'PAM50_' + str(tract_id).zfill(2)] = val * 100.0 / vol_mask_tot
@@ -608,8 +623,8 @@ class AnalyzeLesion:
             ignore_index=True
         )
         self.distrib_matrix_dct[sheet_name].iloc[
-            -1, self.distrib_matrix_dct[sheet_name].columns.get_loc('vert')
-        ] = 'total % (all vert)'
+            -1, self.distrib_matrix_dct[sheet_name].columns.get_loc('row')
+        ] = f'total % (all {self.row_name})'
 
     def __regroup_per_tracts(self, vol_dct, tracts):
         res_mask = [vol_dct[t][0] for t in vol_dct if t in tracts]
@@ -619,28 +634,30 @@ class AnalyzeLesion:
     def _measure_totLesion_distribution(self, im_lesion, atlas_data, im_vert, p_lst, vert_indices):
 
         sheet_name = 'ROI_occupied_by_lesion'
-        self.distrib_matrix_dct[sheet_name] = pd.DataFrame.from_dict({'vert': [str(v) for v in self.vert_lst] + ['total % (all vert)']})
+        self.distrib_matrix_dct[sheet_name] = pd.DataFrame.from_dict({'row': [str(r) for r in self.rows] + [f'total % (all {self.row_name})']})
 
         # initialized to 0 for each vertebral level and each PAM50 tract
         for tract_id in atlas_data:
-            self.distrib_matrix_dct[sheet_name]['PAM50_' + str(tract_id).zfill(2)] = [0] * len(self.vert_lst + ['total % (all vert)'])
+            self.distrib_matrix_dct[sheet_name]['PAM50_' + str(tract_id).zfill(2)] = [0] * len(self.rows + [f'total % (all {self.row_name})'])
 
-        for vert in self.vert_lst + ['total % (all vert)']:  # loop over the vertebral levels
-            print(f"  Computing ROI distribution for vert {vert}...")
-            if vert != 'total % (all vert)':
-                im_vert_cur = self.__keep_only_indices(im_vert, vert_indices[vert])
+        for row in sct_progress_bar(self.rows + [f'total % (all {self.row_name})'], unit=self.row_name,
+                                    desc="  Computing ROI distribution for all lesions"):  # loop over slices/vertlevels
+            if row != f'total % (all {self.row_name})':
+                # Set everything but the current vertlevel (or slice) to 0
+                indices_to_keep = vert_indices[row] if self.row_name == "vert" else row
+                im_vert_cur = self.__keep_only_indices(im_vert, indices_to_keep)
             else:
+                indices_to_keep = None
                 im_vert_cur = None
             if im_vert_cur is None or np.count_nonzero(im_vert_cur * np.copy(im_lesion)):
                 res_perTract_dct = {}  # for each tract compute the volume occupied by lesion and the volume of the tract
-                idx = self.distrib_matrix_dct[sheet_name][self.distrib_matrix_dct[sheet_name].vert == str(vert)].index
+                idx = self.distrib_matrix_dct[sheet_name][self.distrib_matrix_dct[sheet_name].row == str(row)].index
                 for tract_id in atlas_data:  # loop over the tracts
                     res_perTract_dct[tract_id] = self.__relative_ROIvol_in_mask(im_mask_data=np.copy(im_lesion),
                                                                                 im_atlas_roi_data=np.copy(atlas_data[tract_id]),
                                                                                 p_lst=p_lst,
                                                                                 im_template_vert_data=np.copy(im_vert_cur),
-                                                                                vert_level=vert,
-                                                                                vert_indices=vert_indices)
+                                                                                indices_to_keep=indices_to_keep)
 
                 # group tracts to compute involvement in CombinedLabels (GM, WM, DC, VF, LF)
                 for label_name, sublabels in self.atlas_combinedlabels.items():
@@ -666,12 +683,18 @@ class AnalyzeLesion:
             if os.path.isfile(self.path_levels):
                 img_vert = Image(self.path_levels)
                 im_vert_data = img_vert.data
-                self.vert_lst = [v for v in np.unique(im_vert_data) if v]  # list of vertebral levels available in the input image
-                # precompute the list of indices for each vertebral level
-                # these indices are used to zero out certain levels (to compute the volume of the remaining levels)
-                vert_indices = {
-                    vert: np.where(im_vert_data == vert) for vert in self.vert_lst
-                }
+                if self.row_name == "vert":
+                    # list of vertebral levels available in the input image
+                    self.rows = [v for v in np.unique(im_vert_data) if v]
+                    # precompute the list of indices for each vertebral level
+                    # these indices are used to zero out certain levels (to compute the volume of the remaining levels)
+                    vert_indices = {
+                        vert: np.where(im_vert_data == vert) for vert in self.rows
+                    }
+                else:
+                    assert self.row_name == "slice"
+                    self.rows = list(range(im_vert_data.shape[-1]))  # Keep the same vert image, but uses slices instead
+                    vert_indices = None
 
             else:
                 im_vert_data = None
@@ -874,6 +897,7 @@ def main(argv: Sequence[str]):
                                fname_ref=fname_ref,
                                path_template=path_template,
                                path_ofolder=path_results,
+                               perslice=arguments.perslice,
                                verbose=verbose)
 
     # run the analyze