--- a/spinalcordtoolbox/scripts/sct_analyze_lesion.py
+++ b/spinalcordtoolbox/scripts/sct_analyze_lesion.py
@@ -584,7 +584,7 @@ class AnalyzeLesion:
                                                      desc="  Computing lesion distribution (volume values)"):
             # Set everything but the current vertlevel (or slice) to 0
             im_vert_cur = self.__keep_only_indices(im_vert, indices_to_keep)
-            if np.count_nonzero(im_vert_cur * np.copy(im_lesion)):  # if there is lesion in this vertebral level
+            if np.count_nonzero(im_vert_cur * im_lesion):  # if there is lesion in this vertebral level
                 idx = self.distrib_matrix_dct[sheet_name][self.distrib_matrix_dct[sheet_name].row == str(row)].index
                 for tract_id in atlas_data:  # Loop over PAM50 tracts
                     res_lst = self.__relative_ROIvol_in_mask(im_mask_data=np.copy(im_lesion),
@@ -647,7 +647,7 @@ class AnalyzeLesion:
                 im_vert_cur = self.__keep_only_indices(im_vert, indices_to_keep)
             else:
                 im_vert_cur = None
-            if im_vert_cur is None or np.count_nonzero(im_vert_cur * np.copy(im_lesion)):
+            if im_vert_cur is None or np.count_nonzero(im_vert_cur * im_lesion):
                 res_perTract_dct = {}  # for each tract compute the volume occupied by lesion and the volume of the tract
                 idx = self.distrib_matrix_dct[sheet_name][self.distrib_matrix_dct[sheet_name].row == str(row)].index
                 for tract_id in atlas_data:  # loop over the tracts